refactor(utils): relocate report logic
