+++ packages/utils/src/lib/report.spec.ts
@@ -9,7 +9,6 @@ import {
   formatBytes,
   formatCount,
   loadReport,
-  slugify,
 } from './report';
 
 // Mock file system API's
@@ -28,19 +27,6 @@ const resetFiles = (files?: Record<string, string>) => {
   vol.reset();
   vol.fromJSON(files || {}, outputDir);
 };
-
-describe('slugify', () => {
-  it.each([
-    ['Largest Contentful Paint', 'largest-contentful-paint'],
-    ['cumulative-layout-shift', 'cumulative-layout-shift'],
-    ['max-lines-200', 'max-lines-200'],
-    ['rxjs/finnish', 'rxjs-finnish'],
-    ['@typescript-eslint/no-explicit-any', 'typescript-eslint-no-explicit-any'],
-    ['Code  PushUp ', 'code-pushup'],
-  ])('should transform "%s" to valid slug "%s"', (text, slug) => {
-    expect(slugify(text)).toBe(slug);
-  });
-});
 
 describe('formatBytes', () => {
   it('should log file sizes in Bytes`', async () => {
