+++ packages/utils/src/lib/utils.spec.ts
@@ -13,6 +13,7 @@ import {
   toArray,
   toUnixPath,
 } from './utils';
+import { slugify } from './utils';
 
 // Mock file system API's
 vi.mock('fs', async () => {
@@ -25,6 +26,19 @@ vi.mock('fs/promises', async () => {
 });
 
 const outputDir = MEMFS_VOLUME;
+
+describe('slugify', () => {
+  it.each([
+    ['Largest Contentful Paint', 'largest-contentful-paint'],
+    ['cumulative-layout-shift', 'cumulative-layout-shift'],
+    ['max-lines-200', 'max-lines-200'],
+    ['rxjs/finnish', 'rxjs-finnish'],
+    ['@typescript-eslint/no-explicit-any', 'typescript-eslint-no-explicit-any'],
+    ['Code  PushUp ', 'code-pushup'],
+  ])('should transform "%s" to valid slug "%s"', (text, slug) => {
+    expect(slugify(text)).toBe(slug);
+  });
+});
 
 describe('pluralize', () => {
   it.each([
